server/celery-workers/requirements.txt
celery[redis]
python-decouple==3.5
pymongo
dnspython==2.1.0
python-socketio
requests
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.3
future==0.18.3
idna==3.4
joblib==1.2.0
multidict==6.0.4
mutagen==1.46.0
nltk==3.8.1
openai==0.26.1
packaging==23.0
Pillow==9.4.0
pydantic==1.10.4
python-dotenv==0.21.0
regex==2022.10.31
replicate==0.4.0
requests==2.28.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2